refactor: cleanups, renames, add errors
