+++ crates/bridge-proof/protocol2/src/bitcoin_tx.rs
