+++ crates/bridge-proof/protocol2/src/statement.rs
@@ -4,7 +4,8 @@ use strata_proofimpl_btc_blockspace::tx::compute_txid;
 use strata_state::{bridge_state::DepositState, l1::get_btc_params};
 
 use crate::{
-    protocol_op::{extract_checkpoint, extract_claim_info, extract_withdrawal_info},
+    error::BridgeProofError,
+    tx_info::{extract_checkpoint, extract_claim_info, extract_withdrawal_info},
     BridgeProofInput, BridgeProofOutput,
 };
 
@@ -20,6 +21,7 @@ pub fn process_bridge_proof(
             .expect("checkpoint is required");
 
     // 1b. Verify the checkpoint proof is part of the header chain
+    assert!(strata_checkpoint_tx.get_witness_tx().is_some());
     assert!(
         strata_checkpoint_tx.verify(headers[*strata_checkpoint_idx]),
         "invalid checkpoint tx: non-inclusion"
@@ -88,6 +90,7 @@ pub fn process_bridge_proof(
 
     // 6b. Verify that the claim tx is part of the header chain
     let claim_header = headers[*claim_tx_idx];
+    assert!(claim_tx.get_witness_tx().is_some());
     assert!(
         claim_tx.verify(claim_header),
         "invalid claim tx: non-inclusion"
@@ -104,10 +107,12 @@ pub fn process_bridge_proof(
         "invalid claim tx: invalid commitment of anchor idx"
     );
 
+    let headers_after_claim_tx = headers.len() - claim_tx_idx;
     BridgeProofOutput {
         deposit_txid: entry.output().outpoint().txid.into(),
         anchor_key_root: input.anchor_key_root,
         anchor_idx,
         claim_ts: claim_header.time,
+        headers_after_claim_tx,
     }
 }
