+++ crates/bridge-proof/protocol2/src/tx_info.rs
